clickhouse_ecto v0.1.0 ClickhouseEcto.Query

Functions

all(query)

all(query :: Ecto.Query.t()) :: String.t()

Receives a query and must return a SELECT query.

delete(prefix, table, filters, returning)

delete(prefix :: String.t(), table :: String.t(), filters :: [atom()], returning :: [atom()]) :: String.t()

Clickhouse doesn’t support delete

delete_all(query)

delete_all(query :: Ecto.Query.t()) :: String.t()

Clickhouse doesn’t support delete

insert(prefix, table, header, rows, on_conflict, returning)

insert(prefix :: String.t(), table :: String.t(), header :: [atom()], rows :: [[atom() | nil]], on_conflict :: Ecto.Adapter.on_conflict(), returning :: [atom()]) :: String.t()

Returns an INSERT for the given rows in table returning the given returning.

update(prefix, table, fields, filters, returning)

update(prefix :: String.t(), table :: String.t(), fields :: [atom()], filters :: [atom()], returning :: [atom()]) :: String.t()

Clickhouse doesn’t support update

update_all(query, prefix \\ nil)

Receives a query and values to update and must return an UPDATE query.