clickhouse_ecto v0.1.0 ClickhouseEcto.Structure

Functions

structure_dump(default, config)

Dumps the given structure.

The path will be looked in the config under :dump_path or default to the structure path inside default.

Returns :ok if it was loaded successfully, an error tuple otherwise.

Examples

structure_dump("priv/repo", username: "postgres",
                            database: "ecto_test",
                            hostname: "localhost")

Callback implementation for Ecto.Adapter.Structure.structure_dump/2.

structure_load(default, config)

Loads the given structure.

The path will be looked in the config under :dump_path or default to the structure path inside default.

Returns :ok if it was loaded successfully, an error tuple otherwise.

Examples

structure_load("priv/repo", username: "postgres",
                            database: "ecto_test",
                            hostname: "localhost")

Callback implementation for Ecto.Adapter.Structure.structure_load/2.